LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/claim-with-redirect.https-expected.txt

CONSOLE MESSAGE: line 27: Unhandled Promise Rejection: NotSupportedError: ServiceWorkerRegistration::update not yet implemented
  

FAIL Claim works after redirection to another origin assert_equals: expected (string) "updated" but got (undefined) undefined
